new file US_11-03-08-01_trading_psychology_guide_test and modified US_11-03-08_trading_psychology_guide_test
